DRILL-8216: Use EVF-based JSON reader for Values operator

+++ contrib/storage-jdbc/src/test/java/org/apache/drill/exec/store/jdbc/TestJdbcWriterWithPostgres.java
@@ -143,13 +143,13 @@ public class TestJdbcWriterWithPostgres extends ClusterTest {
     DirectRowSet results = queryBuilder().sql(testQuery).rowSet();
 
     TupleMetadata expectedSchema = new SchemaBuilder()
-      .add("ID", MinorType.BIGINT, DataMode.OPTIONAL)
-      .add("NAME", MinorType.BIGINT, DataMode.OPTIONAL)
+      .add("ID", MinorType.INT, DataMode.OPTIONAL)
+      .add("NAME", MinorType.INT, DataMode.OPTIONAL)
       .buildSchema();
 
     RowSet expected = new RowSetBuilder(client.allocator(), expectedSchema)
-      .addRow(1L, 2L)
-      .addRow(3L, 4L)
+      .addRow(1, 2)
+      .addRow(3, 4)
       .build();
 
     RowSetUtilities.verify(expected, results);
@@ -216,13 +216,13 @@ public class TestJdbcWriterWithPostgres extends ClusterTest {
     DirectRowSet results = queryBuilder().sql(testQuery).rowSet();
 
     TupleMetadata expectedSchema = new SchemaBuilder()
-      .add("ID", MinorType.BIGINT, DataMode.OPTIONAL)
-      .add("NAME", MinorType.BIGINT, DataMode.OPTIONAL)
+      .add("ID", MinorType.INT, DataMode.OPTIONAL)
+      .add("NAME", MinorType.INT, DataMode.OPTIONAL)
       .buildSchema();
 
     RowSet expected = new RowSetBuilder(client.allocator(), expectedSchema)
-      .addRow(1L, 2L)
-      .addRow(3L, 4L)
+      .addRow(1, 2)
+      .addRow(3, 4)
       .build();
 
     RowSetUtilities.verify(expected, results);
@@ -297,13 +297,13 @@ public class TestJdbcWriterWithPostgres extends ClusterTest {
     DirectRowSet results = queryBuilder().sql(testQuery).rowSet();
 
     TupleMetadata expectedSchema = new SchemaBuilder()
-      .add("My id", MinorType.BIGINT, DataMode.OPTIONAL)
-      .add("My name", MinorType.BIGINT, DataMode.OPTIONAL)
+      .add("My id", MinorType.INT, DataMode.OPTIONAL)
+      .add("My name", MinorType.INT, DataMode.OPTIONAL)
       .buildSchema();
 
     RowSet expected = new RowSetBuilder(client.allocator(), expectedSchema)
-      .addRow(1L, 2L)
-      .addRow(3L, 4L)
+      .addRow(1, 2)
+      .addRow(3, 4)
       .build();
 
     RowSetUtilities.verify(expected, results);
@@ -366,13 +366,13 @@ public class TestJdbcWriterWithPostgres extends ClusterTest {
     DirectRowSet results = queryBuilder().sql(testQuery).rowSet();
 
     TupleMetadata expectedSchema = new SchemaBuilder()
-      .add("ID", MinorType.BIGINT, DataMode.OPTIONAL)
-      .add("NAME", MinorType.BIGINT, DataMode.OPTIONAL)
+      .add("ID", MinorType.INT, DataMode.OPTIONAL)
+      .add("NAME", MinorType.INT, DataMode.OPTIONAL)
       .buildSchema();
 
     RowSet expected = new RowSetBuilder(client.allocator(), expectedSchema)
-      .addRow(1L, 2L)
-      .addRow(3L, 4L)
+      .addRow(1, 2)
+      .addRow(3, 4)
       .build();
 
     RowSetUtilities.verify(expected, results);
